fix fallback pub key transfer mechanism
The fallback mechanism for transfering the public key to the server used
id_rsa.pub on the client (when it creates id_dsa.pub) and copies it to
id_dsa.pub on the server, instead of appending to authorized_keys as ssh-copy-id
would do.

Use id_dsa.pub and append to authorized_keys.

Signed-off-by: Darren Hart <darren@dvhart.com>